+++ packages/components-js/tests/e2e/specs/components-with-loading-prop.e2e.ts
@@ -9,7 +9,10 @@ import {
   waitForStencilLifecycle,
 } from '../helpers';
 
-const tagNamesWithLoadingProp: TagName[] = TAG_NAMES.filter((tagName) => getComponentMeta(tagName).propsMeta?.loading);
+const tagNamesWithLoadingProp: TagName[] = TAG_NAMES.filter((tagName) => {
+  const meta = getComponentMeta(tagName);
+  return meta.propsMeta?.loading;
+});
 
 for (const tagName of tagNamesWithLoadingProp) {
   test.describe(tagName, () => {
@@ -50,7 +53,7 @@ for (const tagName of tagNamesWithLoadingProp) {
     };
 
     test.describe('for loading="true"', () => {
-      const markup = buildDefaultComponentMarkup(tagName).replace(/>/, ' loading="true">'); // add loading attribute
+      const markup = buildDefaultComponentMarkup(tagName).replace(`<${tagName}`, `<${tagName} loading="true"`); // add loading attribute
 
       test('should render loading message initially', async ({ page }) => {
         await setContentWithDesignSystem(page, markup);