+++ packages/components-react/src/examples/index.ts
@@ -55,6 +55,8 @@ export * from './SegmentedControlExample';
 export * from './SegmentedControlExampleControlled';
 export * from './SelectExample';
 export * from './SelectExampleControlled';
+export * from './RadioGroupExample';
+export * from './RadioGroupExampleControlled';
 export * from './SelectExampleDynamic';
 export * from './SelectExampleRequired';
 export * from './SheetExample';